feat: split requestTab to components and add npmignore settings
