+++ .npmignore
@@ -43,3 +43,6 @@ jspm_packages
 /test
 /.history
 .console.log
+
+server/
+img/
